requirement.txt
streamlit
streamlit-chat
langchain-openai
langchain
langchain-community
python-dotenv
